requirements.txt
tqdm==4.44.0
Flask==1.1.1
Unidecode==1.1.1
pdftotext==2.1.4
pandas==1.0.3
spacy==2.2.4
https://github.com/explosion/spacy-models/releases/download/fr_core_news_md-2.2.5/fr_core_news_md-2.2.5.tar.gz
numpy==1.18.2
Flask_Cors==3.0.8
streamlit==0.57.1
Click==7.1.1
beautifulsoup4==4.8.2
scikit_learn==0.22.2
python-Levenshtein==0.12.0
pytest==5.4.1